fix: file with types declarations
