+++ src/vite-env.d.ts
@@ -1,7 +1,3 @@
-import 'styled-components';
-
-import { AppColors } from './constants';
-
 declare module '*.svg' {
   import React from 'react';
 
@@ -9,12 +5,13 @@ declare module '*.svg' {
   export default SVG;
 }
 
-/// <reference types="vite-plugin-svgr/client" />
-/// <reference types="vite/client" />
-
 declare module '*.jpg';
 declare module '*.png';
 declare module '*.woff';
+
+/// <reference types="vite/client" />
+/// <reference types="vite-plugin-svgr/client" />
+/// <reference types="types/styles/theme" />
 
 interface ImportMetaEnv {
   VITE_APP_TITLE: string;
@@ -29,68 +26,4 @@ interface ImportMetaEnv {
 
 interface ImportMeta {
   readonly env: Readonly<ImportMetaEnv>;
-}
-
-declare module 'styled-components' {
-  export interface DefaultTheme {
-    colors: {
-      main: AppColors;
-      text: AppColors;
-      textAlt: AppColors;
-      bgPrimary: AppColors;
-      bgSecondaryLight: AppColors;
-      bgSecondaryDark: AppColors;
-      error: AppColors;
-    };
-
-    fontFamily: string;
-
-    fontWeight: {
-      rg: number;
-      md: number;
-      bd: number;
-      bk: number;
-    };
-
-    fontSizes: {
-      sm: number;
-      md: number;
-      lg: number;
-      xl: number;
-      xl2: number;
-      xl3: number;
-      xl4: number;
-      xl5: number;
-      xl6: number;
-    };
-
-    breakpoints: {
-      mobileS: number;
-      mobileM: number;
-      mobileL: number;
-      tablet: number;
-      desktopS: number;
-      desktopM: number;
-      desktopL: number;
-    };
-
-    margins: {
-      sm: number;
-      md: number;
-      lg: number;
-      xl: number;
-    };
-
-    radius: {
-      high: number;
-      low: number;
-    };
-
-    opacity: {
-      low: string;
-      high: string;
-    };
-
-    duration: number;
-  }
 }
